Update all dialog components
